remove older code of add/remove server
