+++ svc-aggregation/go.mod
@@ -5,7 +5,7 @@ go 1.13
 require (
 	github.com/ODIM-Project/ODIM/lib-dmtf v0.0.0-20200727115727-33d557ff397c
 	github.com/ODIM-Project/ODIM/lib-messagebus v0.0.0-20200727103018-252e26a63065
-	github.com/ODIM-Project/ODIM/lib-utilities v0.0.0-20200903134721-47bfad56447c
+	github.com/ODIM-Project/ODIM/lib-utilities v0.0.0-20200909041946-ba6d850051fc
 	github.com/ODIM-Project/ODIM/svc-plugin-rest-client v0.0.0-20200727110501-4599893e44fd
 	github.com/satori/go.uuid v1.2.0
 	github.com/stretchr/testify v1.5.1
