+++ svc-aggregation/rpc/aggregator.go
@@ -126,193 +126,6 @@ func validateManagerAddress(managerAddress string) error {
 	return nil
 }
 
-// AddCompute defines the operations which handles the RPC request response
-// for the create account service of aggregator micro service.
-// The functionality retrives the request and return backs the response to
-// RPC according to the protoc file defined in the util-lib package.
-// The function also checks for the session time out of the token
-// which is present in the request.
-func (a *Aggregator) AddCompute(ctx context.Context, req *aggregatorproto.AggregatorRequest, resp *aggregatorproto.AggregatorResponse) error {
-	var taskID string
-	var oemprivileges []string
-	privileges := []string{common.PrivilegeConfigureComponents}
-	authStatusCode, authStatusMessage := a.connector.Auth(req.SessionToken, privileges, oemprivileges)
-	if authStatusCode != http.StatusOK {
-		errMsg := "error while trying to authenticate session"
-		generateResponse(common.GeneralError(authStatusCode, authStatusMessage, errMsg, nil, nil), resp)
-		log.Printf(errMsg)
-		return nil
-	}
-	sessionUserName, err := a.connector.GetSessionUserName(req.SessionToken)
-	if err != nil {
-		errMsg := "error while trying to get the session username: " + err.Error()
-		generateResponse(common.GeneralError(http.StatusUnauthorized, response.NoValidSession, errMsg, nil, nil), resp)
-		log.Printf(errMsg)
-		return nil
-	}
-
-	// parsing the AddResourceRequest
-	var addResourceRequest system.AddResourceRequest
-	err = json.Unmarshal(req.RequestBody, &addResourceRequest)
-	if err != nil {
-		errMsg := "unable to parse the add request" + err.Error()
-		generateResponse(common.GeneralError(http.StatusInternalServerError, response.InternalError, errMsg, nil, nil), resp)
-		log.Printf(errMsg)
-		return nil
-	}
-
-	//validating the AddResourceRequest
-	invalidParam := validateAddResoureRequest(addResourceRequest)
-	if invalidParam != "" {
-		errMsg := "error: Mandatory field " + invalidParam + " Missing"
-		generateResponse(common.GeneralError(http.StatusBadRequest, response.PropertyMissing, errMsg, []interface{}{invalidParam}, nil), resp)
-		log.Printf(errMsg)
-		return nil
-	}
-	managerAddress := addResourceRequest.ManagerAddress
-	err = validateManagerAddress(managerAddress)
-	if err != nil {
-		generateResponse(common.GeneralError(http.StatusBadRequest, response.PropertyValueFormatError, err.Error(), []interface{}{managerAddress, "ManagerAddress"}, nil), resp)
-		log.Printf(err.Error())
-		return nil
-	}
-
-	// Task Service using RPC and get the taskID
-	taskURI, err := a.connector.CreateTask(sessionUserName)
-	if err != nil {
-		errMsg := "error while trying to create the task: " + err.Error()
-		generateResponse(common.GeneralError(http.StatusInternalServerError, response.InternalError, errMsg, nil, nil), resp)
-		log.Printf(errMsg)
-		return nil
-	}
-	strArray := strings.Split(taskURI, "/")
-	if strings.HasSuffix(taskURI, "/") {
-		taskID = strArray[len(strArray)-2]
-	} else {
-		taskID = strArray[len(strArray)-1]
-	}
-	// spawn the thread here to process the action asynchronously
-	go a.connector.AggregationServiceAdd(taskID, sessionUserName, req)
-
-	// return 202 Accepted
-	var rpcResp = response.RPC{
-		StatusCode:    http.StatusAccepted,
-		StatusMessage: response.TaskStarted,
-		Header: map[string]string{
-			"Content-type": "application/json; charset=utf-8",
-			"Location":     "/taskmon/" + taskID,
-		},
-	}
-	generateTaskRespone(taskID, taskURI, &rpcResp)
-	generateResponse(rpcResp, resp)
-
-	return nil
-}
-
-// DeleteCompute defines the operations which handles the RPC request response
-// for the delete compute system of aggregator micro service.
-// The functionality retrives the request and delete the compute system and return backs the response to
-// RPC according to the protoc file defined in the util-lib package.
-// The function also checks for the session time out of the token
-// which is present in the request.
-func (a *Aggregator) DeleteCompute(ctx context.Context, req *aggregatorproto.AggregatorRequest, resp *aggregatorproto.AggregatorResponse) error {
-	var oemprivileges []string
-	// Task Service using RPC and get the taskID
-	targetURI := "/redfish/v1/AggregationService/Actions/AggregationService.Delete"
-	privileges := []string{common.PrivilegeConfigureComponents}
-	authStatusCode, authStatusMessage := a.connector.Auth(req.SessionToken, privileges, oemprivileges)
-	if authStatusCode != http.StatusOK {
-		errMsg := "error while trying to authenticate session"
-		generateResponse(common.GeneralError(authStatusCode, authStatusMessage, errMsg, nil, nil), resp)
-		log.Printf(errMsg)
-		return nil
-	}
-	sessionUserName, err := a.connector.GetSessionUserName(req.SessionToken)
-	if err != nil {
-		errMsg := "error while trying to get the session username: " + err.Error()
-		generateResponse(common.GeneralError(http.StatusUnauthorized, response.NoValidSession, errMsg, nil, nil), resp)
-		log.Printf(errMsg)
-		return nil
-	}
-
-	// Task Service using RPC and get the taskID
-	taskURI, err := a.connector.CreateTask(sessionUserName)
-	if err != nil {
-		errMsg := "error while trying to create the task: " + err.Error()
-		generateResponse(common.GeneralError(http.StatusInternalServerError, response.InternalError, errMsg, nil, nil), resp)
-		log.Printf(errMsg)
-		return nil
-	}
-	var taskID string
-	strArray := strings.Split(taskURI, "/")
-	if strings.HasSuffix(taskURI, "/") {
-		taskID = strArray[len(strArray)-2]
-	} else {
-		taskID = strArray[len(strArray)-1]
-	}
-	go deleteServer(taskID, targetURI, a, req)
-	// return 202 Accepted
-	var rpcResp = response.RPC{
-		StatusCode:    http.StatusAccepted,
-		StatusMessage: response.TaskStarted,
-		Header: map[string]string{
-			"Content-type": "application/json; charset=utf-8",
-			"Location":     "/taskmon/" + taskID,
-		},
-	}
-	generateTaskRespone(taskID, taskURI, &rpcResp)
-	generateResponse(rpcResp, resp)
-
-	return nil
-
-}
-
-func deleteServer(taskID string, targetURI string, a *Aggregator, req *aggregatorproto.AggregatorRequest) error {
-	err := a.connector.UpdateTask(common.TaskData{
-		TaskID:          taskID,
-		TargetURI:       targetURI,
-		TaskState:       common.Running,
-		TaskStatus:      common.OK,
-		PercentComplete: 0,
-		HTTPMethod:      http.MethodPost,
-	})
-	if err != nil && (err.Error() == common.Cancelling) {
-		// We cant do anything here as the task has done it work completely, we cant reverse it.
-		//Unless if we can do opposite/reverse action for delete server which is add server.
-		a.connector.UpdateTask(common.TaskData{
-			TaskID:          taskID,
-			TargetURI:       targetURI,
-			TaskState:       common.Cancelled,
-			TaskStatus:      common.OK,
-			PercentComplete: 0,
-			HTTPMethod:      http.MethodPost,
-		})
-		go runtime.Goexit()
-	}
-	data := a.connector.DeleteCompute(req)
-	err = a.connector.UpdateTask(common.TaskData{
-		TaskID:          taskID,
-		TargetURI:       targetURI,
-		TaskState:       common.Completed,
-		TaskStatus:      common.OK,
-		Response:        data,
-		PercentComplete: 100,
-		HTTPMethod:      http.MethodPost,
-	})
-	if err != nil && (err.Error() == common.Cancelling) {
-		a.connector.UpdateTask(common.TaskData{
-			TaskID:          taskID,
-			TargetURI:       targetURI,
-			TaskState:       common.Cancelled,
-			TaskStatus:      common.OK,
-			PercentComplete: 100,
-			HTTPMethod:      http.MethodPost,
-		})
-		go runtime.Goexit()
-	}
-	return nil
-}
-
 // Reset function is for handling the RPC communication for Reset Action
 func (a *Aggregator) Reset(ctx context.Context, req *aggregatorproto.AggregatorRequest, resp *aggregatorproto.AggregatorResponse) error {
