+++ svc-aggregation/rpc/aggregator_test.go
@@ -82,251 +82,6 @@ func TestAggregator_GetAggregationService(t *testing.T) {
 	}
 }
 
-func TestAggregator_AddCompute(t *testing.T) {
-	config.SetUpMockConfig(t)
-	addComputeRetrieval := config.AddComputeSkipResources{
-		SystemCollection: []string{"Chassis", "LogServices"},
-	}
-	mockPluginData(t, "ILO")
-
-	config.Data.AddComputeSkipResources = &addComputeRetrieval
-	system.ActiveReqSet.ReqRecord = make(map[string]interface{})
-	defer func() {
-		common.TruncateDB(common.OnDisk)
-		common.TruncateDB(common.InMemory)
-	}()
-
-	successReq, _ := json.Marshal(system.AddResourceRequest{
-		ManagerAddress: "100.0.0.1:50000",
-		UserName:       "admin",
-		Password:       "password",
-		Oem: &system.AddOEM{
-			PluginID:          "GRF",
-			PreferredAuthType: "BasicAuth",
-			PluginType:        "RF-GENERIC",
-		},
-	})
-	invalidReqBody, _ := json.Marshal(system.AddResourceRequest{
-		ManagerAddress: ":50000",
-		UserName:       "admin",
-		Password:       "password",
-		Oem: &system.AddOEM{
-			PluginID:          "GRF",
-			PreferredAuthType: "BasicAuth",
-			PluginType:        "RF-GENERIC",
-		},
-	})
-	missingparamReq, _ := json.Marshal(system.AddResourceRequest{})
-	type args struct {
-		ctx  context.Context
-		req  *aggregatorproto.AggregatorRequest
-		resp *aggregatorproto.AggregatorResponse
-	}
-	tests := []struct {
-		name    string
-		a       *Aggregator
-		args    args
-		wantErr bool
-	}{
-		{
-			name: "positive case",
-			a:    &Aggregator{connector: connector},
-			args: args{
-				req:  &aggregatorproto.AggregatorRequest{SessionToken: "validToken", RequestBody: successReq},
-				resp: &aggregatorproto.AggregatorResponse{},
-			},
-			wantErr: false,
-		},
-		{
-			name: "auth fail",
-			a:    &Aggregator{connector: connector},
-			args: args{
-				req:  &aggregatorproto.AggregatorRequest{SessionToken: "invalidToken", RequestBody: successReq},
-				resp: &aggregatorproto.AggregatorResponse{},
-			},
-			wantErr: false,
-		},
-		{
-			name: "get session username fails",
-			a:    &Aggregator{connector: connector},
-			args: args{
-				req:  &aggregatorproto.AggregatorRequest{SessionToken: "noDetailsToken", RequestBody: successReq},
-				resp: &aggregatorproto.AggregatorResponse{},
-			},
-			wantErr: false,
-		},
-		{
-			name: "unable to create task",
-			a:    &Aggregator{connector: connector},
-			args: args{
-				req:  &aggregatorproto.AggregatorRequest{SessionToken: "noTaskToken", RequestBody: successReq},
-				resp: &aggregatorproto.AggregatorResponse{},
-			},
-			wantErr: false,
-		},
-		{
-			name: "task with slash",
-			a:    &Aggregator{connector: connector},
-			args: args{
-				req:  &aggregatorproto.AggregatorRequest{SessionToken: "taskWithSlashToken", RequestBody: successReq},
-				resp: &aggregatorproto.AggregatorResponse{},
-			},
-			wantErr: false,
-		},
-		{
-			name: "with invalid request",
-			a:    &Aggregator{connector: connector},
-			args: args{
-				req:  &aggregatorproto.AggregatorRequest{SessionToken: "validToken", RequestBody: []byte("someData")},
-				resp: &aggregatorproto.AggregatorResponse{},
-			},
-			wantErr: false,
-		},
-		{
-			name: "Invalid Manager Address",
-			a:    &Aggregator{connector: connector},
-			args: args{
-				req:  &aggregatorproto.AggregatorRequest{SessionToken: "validToken", RequestBody: invalidReqBody},
-				resp: &aggregatorproto.AggregatorResponse{},
-			},
-			wantErr: false,
-		},
-		{
-			name: "with missing parameters",
-			a:    &Aggregator{connector: connector},
-			args: args{
-				req:  &aggregatorproto.AggregatorRequest{SessionToken: "validToken", RequestBody: missingparamReq},
-				resp: &aggregatorproto.AggregatorResponse{},
-			},
-			wantErr: false,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if err := tt.a.AddCompute(tt.args.ctx, tt.args.req, tt.args.resp); (err != nil) != tt.wantErr {
-				t.Errorf("Aggregator.AddCompute() error = %v, wantErr %v", err, tt.wantErr)
-			}
-		})
-	}
-}
-
-func TestAggregator_DeleteCompute(t *testing.T) {
-	successReq, _ := json.Marshal(`map[string]interface{}{"parameters": []Parameters{{Name: "/redfish/v1/systems/ef83e569-7336-492a-aaee-31c02d9db831:1"}}}`)
-	type args struct {
-		ctx  context.Context
-		req  *aggregatorproto.AggregatorRequest
-		resp *aggregatorproto.AggregatorResponse
-	}
-	tests := []struct {
-		name    string
-		a       *Aggregator
-		args    args
-		wantErr bool
-	}{
-		{
-			name: "positive case",
-			a:    &Aggregator{connector: connector},
-			args: args{
-				req:  &aggregatorproto.AggregatorRequest{SessionToken: "validToken", RequestBody: successReq},
-				resp: &aggregatorproto.AggregatorResponse{},
-			},
-			wantErr: false,
-		},
-
-		{
-			name: "auth fail",
-			a:    &Aggregator{connector: connector},
-			args: args{
-				req:  &aggregatorproto.AggregatorRequest{SessionToken: "invalidToken", RequestBody: successReq},
-				resp: &aggregatorproto.AggregatorResponse{},
-			},
-			wantErr: false,
-		},
-		{
-			name: "get session username fails",
-			a:    &Aggregator{connector: connector},
-			args: args{
-				req:  &aggregatorproto.AggregatorRequest{SessionToken: "noDetailsToken", RequestBody: successReq},
-				resp: &aggregatorproto.AggregatorResponse{},
-			},
-			wantErr: false,
-		},
-		{
-			name: "unable to create task",
-			a:    &Aggregator{connector: connector},
-			args: args{
-				req:  &aggregatorproto.AggregatorRequest{SessionToken: "noTaskToken", RequestBody: successReq},
-				resp: &aggregatorproto.AggregatorResponse{},
-			},
-			wantErr: false,
-		},
-		{
-			name: "task with slash",
-			a:    &Aggregator{connector: connector},
-			args: args{
-				req:  &aggregatorproto.AggregatorRequest{SessionToken: "taskWithSlashToken", RequestBody: successReq},
-				resp: &aggregatorproto.AggregatorResponse{},
-			},
-			wantErr: false,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if err := tt.a.DeleteCompute(tt.args.ctx, tt.args.req, tt.args.resp); (err != nil) != tt.wantErr {
-				t.Errorf("Aggregator.DeleteCompute() error = %v, wantErr %v", err, tt.wantErr)
-			}
-		})
-	}
-}
-
-func Test_deleteServer(t *testing.T) {
-	type args struct {
-		taskID    string
-		targetURI string
-		a         *Aggregator
-		req       *aggregatorproto.AggregatorRequest
-	}
-	tests := []struct {
-		name    string
-		args    args
-		wantErr bool
-	}{
-		{
-			name: "positive case",
-			args: args{
-				taskID:    "someID",
-				targetURI: "someURI",
-				a:         &Aggregator{connector: connector},
-				req: &aggregatorproto.AggregatorRequest{
-					SessionToken: "validToken",
-					RequestBody:  []byte("someData"),
-				},
-			},
-			wantErr: false,
-		},
-		{
-			name: "task updation fails",
-			args: args{
-				taskID:    "invalid",
-				targetURI: "someURI",
-				a:         &Aggregator{connector: connector},
-				req: &aggregatorproto.AggregatorRequest{
-					SessionToken: "validToken",
-					RequestBody:  []byte("successReq"),
-				},
-			},
-			wantErr: false,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if err := deleteServer(tt.args.taskID, tt.args.targetURI, tt.args.a, tt.args.req); (err != nil) != tt.wantErr {
-				t.Errorf("deleteServer() error = %v, wantErr %v", err, tt.wantErr)
-			}
-		})
-	}
-}
-
 func TestAggregator_Reset(t *testing.T) {
 	successReq, _ := json.Marshal(`map[string]interface{}{"parameters": []Parameters{{Name: "/redfish/v1/systems/ef83e569-7336-492a-aaee-31c02d9db831:1"}}}`)
 	type args struct {
