+++ svc-aggregation/system/addaggregationsource_test.go
@@ -19,6 +19,10 @@ import (
 	"net/http"
 	"reflect"
 	"testing"
+ "strings"
+ "io/ioutil"
+ "fmt"
+ "bytes"
 	"time"
 
 	"github.com/ODIM-Project/ODIM/lib-utilities/common"
@@ -27,6 +31,224 @@ import (
 	"github.com/ODIM-Project/ODIM/lib-utilities/response"
 	"github.com/ODIM-Project/ODIM/svc-aggregation/agmodel"
 )
+
+func EventFunctionsForTesting(s []string)                 {}
+func PostEventFunctionForTesting(s []string, name string) {}
+func GetPluginStatusForTesting(plugin agmodel.Plugin) bool {
+	return true
+}
+func mockSubscribeEMB(pluginID string, list []string) {
+	return
+}
+
+func testContactClientWithDelay(url, method, token string, odataID string, body interface{}, credentials map[string]string) (*http.Response, error) {
+	time.Sleep(4 * time.Second)
+	if strings.Contains(url, "/ODIM/v1/Systems/1") {
+		uid := "24b243cf-f1e3-5318-92d9-2d6737d6b0b9"
+		body := `{"@odata.id":"/ODIM/v1/Systems/1", "UUID": "` + uid + `", "Id": "1",
+		    "MemorySummary":{
+			"Status":{
+			"HealthRollup": "OK"
+			},
+			"TotalSystemMemoryGiB": 384,
+			"TotalSystemPersistentMemoryGiB": 0
+			},
+			"PowerState": "On",
+			"ProcessorSummary":{
+				"Count": 2,
+				"Model": "Intel(R) Xeon(R) Gold 6152 CPU @ 2.10GHz",
+				"Status":{
+					"HealthRollup": "OK"
+				}
+			},
+			"SystemType": "Physical",
+		}`
+		return &http.Response{
+			StatusCode: http.StatusOK,
+			Body:       ioutil.NopCloser(bytes.NewBufferString(body)),
+		}, nil
+	}
+	fBody := `{"Members":[{"@odata.id":"/ODIM/v1/Systems/1"}]}`
+	return &http.Response{
+		StatusCode: http.StatusOK,
+		Body:       ioutil.NopCloser(bytes.NewBufferString(fBody)),
+	}, nil
+}
+
+func mockContactClientForDuplicate(url, method, token string, odataID string, body interface{}, credentials map[string]string) (*http.Response, error) {
+	var bData agmodel.SaveSystem
+	bBytes, _ := json.Marshal(body)
+	json.Unmarshal(bBytes, &bData)
+	host := strings.Split(url, "/ODIM")[0]
+	uid := "test1"
+	if url == "https://localhost:9091/ODIM/v1/Systems/1/Actions/ComputerSystem.Add" {
+		body := `{"MessageId": "Base.1.0.Success"}`
+		return &http.Response{
+			StatusCode: http.StatusOK,
+			Body:       ioutil.NopCloser(bytes.NewBufferString(body)),
+		}, nil
+	} else if url == "https://localhost:9091/ODIM/v1/Registries" {
+		body := `{"Members":[{"@odata.id":"/redfish/v1/Registries/Base"},{"@odata.id":"/redfish/v1/Registries/SomeMember"}]}`
+		return &http.Response{
+			StatusCode: http.StatusOK,
+			Body:       ioutil.NopCloser(bytes.NewBufferString(body)),
+		}, nil
+	} else if url == "https://localhost:9091/ODIM/v1/Registries/Base" {
+		body := `{"@odata.context":"/redfish/v1/$metadata#MessageRegistryFile.MessageRegistryFile","@odata.etag":"W/\"0DCA67A0\"","@odata.id":"/redfish/v1/Registries/Base","@odata.type":"#MessageRegistryFile.v1_0_4.MessageRegistryFile","Id":"Base","Description":"Registry Definition File for Base","Languages":["en"],"Location":[{"Language":"en","Uri":"/redfish/v1/RegistryStore/registries/en/Base.json"}],"Name":"Base Message Registry File","Registry":"Base.1.4.0"}`
+		return &http.Response{
+			StatusCode: http.StatusOK,
+			Body:       ioutil.NopCloser(bytes.NewBufferString(body)),
+		}, nil
+	} else if url == "https://localhost:9091/ODIM/v1/Registries/SomeMember" {
+		body := `{"@odata.context":"/redfish/v1/$metadata#MessageRegistryFile.MessageRegistryFile","@odata.etag":"W/\"0DCA67A0\"","@odata.id":"/redfish/v1/Registries/Base","@odata.type":"#MessageRegistryFile.v1_0_4.MessageRegistryFile","Id":"Base","Description":"Registry Definition File for Base","Languages":["en"],"Location":[{"Language":"en","Uri":"/redfish/v1/RegistryStore/registries/en/SomeRegistry.json"}],"Name":"Base Message Registry File","Registry":"SomeRegistry"}`
+		return &http.Response{
+			StatusCode: http.StatusOK,
+			Body:       ioutil.NopCloser(bytes.NewBufferString(body)),
+		}, nil
+	} else if strings.Contains(url, "SomeRegistry.json") {
+		body := `{"MessageId": "Base.1.0.Success"}`
+		return &http.Response{
+			StatusCode: http.StatusOK,
+			Body:       ioutil.NopCloser(bytes.NewBufferString(body)),
+		}, nil
+	} else if url == "https://localhost:9091/ODIM/v1/Systems" {
+		body := `{"Members":[{"@odata.id":"/ODIM/v1/Systems/1"}]}`
+		return &http.Response{
+			StatusCode: http.StatusOK,
+			Body:       ioutil.NopCloser(bytes.NewBufferString(body)),
+		}, nil
+
+	} else if url == "https://localhost:9091/ODIM/v1/Systems/1" {
+		body := `{"@odata.id":"/ODIM/v1/Systems/1", "UUID": "` + uid + `", "Id": "1",
+		    "MemorySummary":{
+			"Status":{
+			"HealthRollup": "OK"
+			},
+			"TotalSystemMemoryGiB": 384,
+			"TotalSystemPersistentMemoryGiB": 0
+			},
+			"PowerState": "On",
+"ProcessorSummary":{
+"Count": 2,
+"Model": "Intel(R) Xeon(R) Gold 6152 CPU @ 2.10GHz",
+"Status":{
+"HealthRollup": "OK"
+}
+},"SystemType": "Physical",
+  "Links":{
+	"ManagedBy":[
+	{
+	"@odata.id": "/redfish/v1/Managers/1"
+	}
+	]
+	},
+	"Storage":{
+		"@odata.id": "/redfish/v1/Systems/1/Storage"
+		}}`
+		return &http.Response{
+			StatusCode: http.StatusOK,
+			Body:       ioutil.NopCloser(bytes.NewBufferString(body)),
+		}, nil
+
+	} else if url == "https://localhost:9091/ODIM/v1/Chassis" {
+		body := `{"Members":[{"@odata.id":"/ODIM/v1/Chassis/1"}]}`
+		return &http.Response{
+			StatusCode: http.StatusOK,
+			Body:       ioutil.NopCloser(bytes.NewBufferString(body)),
+		}, nil
+
+	} else if url == "https://localhost:9091/ODIM/v1/Chassis/1" {
+		body := `{"@odata.id":"/ODIM/v1/Chassis/1", "UUID": "` + uid + `", "Id": "1"}`
+		return &http.Response{
+			StatusCode: http.StatusOK,
+			Body:       ioutil.NopCloser(bytes.NewBufferString(body)),
+		}, nil
+
+	} else if url == "https://localhost:9091/ODIM/v1/Managers/1" {
+		body := `{"@odata.id":"/ODIM/v1/Managers/1", "UUID": "1s7sda8asd-asdas8as0", "Id": "1","FirmwareVersion": "iLO 5 v2.12"}`
+		return &http.Response{
+			StatusCode: http.StatusOK,
+			Body:       ioutil.NopCloser(bytes.NewBufferString(body)),
+		}, nil
+
+	} else if url == "https://localhost:9091/ODIM/v1/Systems/1/Storage" {
+		body := `{"Members":[{"@odata.id":"/ODIM/v1/Systems/1/Storage/1"}]}`
+		return &http.Response{
+			StatusCode: http.StatusOK,
+			Body:       ioutil.NopCloser(bytes.NewBufferString(body)),
+		}, nil
+
+	} else if url == "https://localhost:9091/ODIM/v1/Systems/1/Storage/1" {
+		body := `{"Drives":[
+			{
+			"@odata.id": "/ODIM/v1/Systems/1/Storage/1/Drives/0"
+			}
+			]}`
+		return &http.Response{
+			StatusCode: http.StatusOK,
+			Body:       ioutil.NopCloser(bytes.NewBufferString(body)),
+		}, nil
+
+	} else if url == "https://localhost:9091/ODIM/v1/Systems/1/Storage/1/Drives/0" {
+		body := `{"BlockSizeBytes": 512,
+		"CapacityBytes": 2179699264,
+		"MediaType": "HDD"}`
+		return &http.Response{
+			StatusCode: http.StatusOK,
+			Body:       ioutil.NopCloser(bytes.NewBufferString(body)),
+		}, nil
+
+	} else if url == host+"/ODIM/v1/Managers" {
+		body := `{"Members":[{"@odata.id":"/ODIM/v1/Managers/1"}]}`
+		if host == "https://100.0.0.5:9091" {
+			return nil, fmt.Errorf("manager data not available not reachable")
+		}
+		if host == "https://100.0.0.6:9091" {
+			body = "incorrectResponse"
+		}
+
+		return &http.Response{
+			StatusCode: http.StatusOK,
+			Body:       ioutil.NopCloser(bytes.NewBufferString(body)),
+		}, nil
+
+	} else if url == host+"/ODIM/v1/Managers/1" {
+		body := `{"@odata.id":"/ODIM/v1/Managers/1", "UUID": "1s7sda8asd-asdas8as0", "Id": "1"}`
+		return &http.Response{
+			StatusCode: http.StatusOK,
+			Body:       ioutil.NopCloser(bytes.NewBufferString(body)),
+		}, nil
+
+	} else if url == host+"/ODIM/v1/Status" {
+		body := `{"EventMessageBus":{"EmbQueue":[{"EmbQueueName":"GRF"}]}}`
+		if host == "https://100.0.0.3:9091" {
+			return nil, fmt.Errorf("plugin not reachable")
+		}
+		if host == "https://100.0.0.4:9091" {
+			body = "incorrectResponse"
+		}
+		return &http.Response{
+			StatusCode: http.StatusOK,
+			Body:       ioutil.NopCloser(bytes.NewBufferString(body)),
+		}, nil
+
+	} else if strings.Contains(url, "/ODIM/v1/validate") || url == "https://localhost:9091/ODIM/v1/Sessions" || url == host+"/ODIM/v1/Sessions" {
+		body := `{"MessageId": "Base.1.0.Success"}`
+		if bData.UserName == "incorrectusername" || bytes.Compare(bData.Password, []byte("incorrectPassword")) == 0 {
+			return &http.Response{
+				StatusCode: http.StatusUnauthorized,
+				Body:       ioutil.NopCloser(bytes.NewBufferString(body)),
+			}, nil
+		}
+		return &http.Response{
+			StatusCode: http.StatusOK,
+			Body:       ioutil.NopCloser(bytes.NewBufferString(body)),
+		}, nil
+
+	}
+
+	return nil, fmt.Errorf("InvalidRequest")
+}
 
 func TestExternalInterface_AddBMC(t *testing.T) {
 	common.MuxLock.Lock()
