+++ svc-api/handle/aggregator.go
@@ -29,8 +29,6 @@ import (
 // AggregatorRPCs defines all the RPC methods in aggregator service
 type AggregatorRPCs struct {
 	GetAggregationServiceRPC                func(aggregatorproto.AggregatorRequest) (*aggregatorproto.AggregatorResponse, error)
-	AddComputeRPC                           func(aggregatorproto.AggregatorRequest) (*aggregatorproto.AggregatorResponse, error)
-	DeleteComputeRPC                        func(aggregatorproto.AggregatorRequest) (*aggregatorproto.AggregatorResponse, error)
 	ResetRPC                                func(aggregatorproto.AggregatorRequest) (*aggregatorproto.AggregatorResponse, error)
 	SetDefaultBootOrderRPC                  func(aggregatorproto.AggregatorRequest) (*aggregatorproto.AggregatorResponse, error)
 	AddAggregationSourceRPC                 func(aggregatorproto.AggregatorRequest) (*aggregatorproto.AggregatorResponse, error)
@@ -75,110 +73,6 @@ func (a *AggregatorRPCs) GetAggregationService(ctx iris.Context) {
 	ctx.StatusCode(int(resp.StatusCode))
 	ctx.Write(resp.Body)
 
-}
-
-// AddCompute is the handler for adding system
-func (a *AggregatorRPCs) AddCompute(ctx iris.Context) {
-	var req interface{}
-	err := ctx.ReadJSON(&req)
-	if err != nil {
-		errorMessage := "error while trying to get JSON body from the aggregator request body: " + err.Error()
-		log.Println(errorMessage)
-		response := common.GeneralError(http.StatusBadRequest, response.MalformedJSON, errorMessage, nil, nil)
-		ctx.StatusCode(http.StatusBadRequest) // TODO: add error headers
-		ctx.JSON(&response.Body)
-		return
-	}
-
-	sessionToken := ctx.Request().Header.Get("X-Auth-Token")
-
-	if sessionToken == "" {
-		errorMessage := "error: no X-Auth-Token found in request header"
-		log.Println(errorMessage)
-		response := common.GeneralError(http.StatusUnauthorized, response.NoValidSession, errorMessage, nil, nil)
-		ctx.StatusCode(http.StatusUnauthorized) // TODO: add error headers
-		ctx.JSON(&response.Body)
-		return
-	}
-
-	// marshalling the req to make aggregator add request
-	// Since aggregator add request accepts []byte stream
-	request, err := json.Marshal(req)
-
-	addRequest := aggregatorproto.AggregatorRequest{
-		SessionToken: sessionToken,
-		RequestBody:  request,
-	}
-	resp, err := a.AddComputeRPC(addRequest)
-	if err != nil {
-		errorMessage := "error: something went wrong with the RPC calls: " + err.Error()
-		log.Println(errorMessage)
-		response := common.GeneralError(http.StatusInternalServerError, response.InternalError, errorMessage, nil, nil)
-		ctx.StatusCode(http.StatusInternalServerError) // TODO: add error headers
-		ctx.JSON(&response.Body)
-		return
-	}
-
-	common.SetResponseHeader(ctx, resp.Header)
-	ctx.StatusCode(int(resp.StatusCode))
-	ctx.Write(resp.Body)
-}
-
-// DeleteCompute is the handler for deleting compute system
-// from iris context will get the request and check sessiontoken
-// and do rpc call and send response back
-func (a *AggregatorRPCs) DeleteCompute(ctx iris.Context) {
-	var req interface{}
-	err := ctx.ReadJSON(&req)
-	if err != nil {
-		errorMessage := "error while trying to get JSON body from the aggregator request body: " + err.Error()
-		log.Println(errorMessage)
-		response := common.GeneralError(http.StatusBadRequest, response.MalformedJSON, errorMessage, nil, nil)
-		ctx.StatusCode(http.StatusBadRequest) // TODO: add error headers
-		ctx.JSON(&response.Body)
-		return
-	}
-
-	sessionToken := ctx.Request().Header.Get("X-Auth-Token")
-
-	if sessionToken == "" {
-		errorMessage := "error: no X-Auth-Token found in request header"
-		log.Println(errorMessage)
-		response := common.GeneralError(http.StatusUnauthorized, response.NoValidSession, errorMessage, nil, nil)
-		ctx.StatusCode(http.StatusUnauthorized) // TODO: add error headers
-		ctx.JSON(&response.Body)
-		return
-	}
-
-	// marshalling the req to make aggregator delete request
-	// Since aggregator deleterequest accepts []byte stream
-	request, err := json.Marshal(req)
-	if err != nil {
-		errorMessage := "error while trying to create JSON request body: " + err.Error()
-		log.Println(errorMessage)
-		response := common.GeneralError(http.StatusInternalServerError, response.InternalError, errorMessage, nil, nil)
-		ctx.StatusCode(http.StatusInternalServerError) // TODO: add error headers
-		ctx.JSON(&response.Body)
-		return
-	}
-
-	deleteRequest := aggregatorproto.AggregatorRequest{
-		SessionToken: sessionToken,
-		RequestBody:  request,
-	}
-	resp, err := a.DeleteComputeRPC(deleteRequest)
-	if err != nil {
-		errorMessage := "error: RPC error: " + err.Error()
-		log.Println(errorMessage)
-		response := common.GeneralError(http.StatusInternalServerError, response.InternalError, errorMessage, nil, nil)
-		ctx.StatusCode(http.StatusInternalServerError) // TODO: add error headers
-		ctx.JSON(&response.Body)
-		return
-	}
-
-	common.SetResponseHeader(ctx, resp.Header)
-	ctx.StatusCode(int(resp.StatusCode))
-	ctx.Write(resp.Body)
 }
 
 // Reset is the handler to reset compute system
